backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.62",
    "fastapi>=0.135.1",
    "numpy>=2.0.0",
    "pillow>=11.0.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.32.5",
    "routes>=2.5.1",
    "shapely>=2.0.0",
    "torch>=2.0.0",
    "uvicorn[standard]>=0.38.0",

]
